add firstinst-gconf-nautilus-no-autorun package
Disable automount and autorun in Nautilus. When autorun and automount
were enabled, Nautilus was mounting /home and other partitions and
popping up windows of them during the post-install step of installation
of the OS to the hard drive. This may have generated exceptions in
anaconda when an attempt to unmount them was made.